+++ Changelog.md
@@ -27,6 +27,7 @@
 - Add new states to the license store - [#763](https://github.com/PrefectHQ/ui/pull/763)
 - Add a tooltip for task run names in the flow run task table - [#814](https://github.com/PrefectHQ/ui/pull/814)
 - Add "Task" to the "Runs Today" to make it clear it's "Task runs today" - [#816](https://github.com/PrefectHQ/ui/pull/816)
+- Add cancel all and work queue paused system actions, VU for clear late runs on the dashboard - [#825](https://github.com/PrefectHQ/ui/pull/825)
 
 ### Bugfixes